refactor: move crit_section.h to arm_generic dir
